icart_backend/requirements.txt
alabaster==0.7.16
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
Babel==2.14.0
bcrypt==4.1.2
billiard==4.2.0
binaryornot==0.4.4
boto3==1.34.49
botocore==1.34.49
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Collectfast==2.2.0
colorama==0.4.6
cookiecutter==2.6.0
coverage==7.4.3
crispy-bootstrap5==2023.10
cron-descriptor==1.4.3
cryptography==42.0.5
cssbeautifier==1.15.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.10
django-allauth==0.61.1
django-anymail==10.2
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-coverage-plugin==3.1.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==4.4.0
django-ninja==1.1.0
django-redis==5.4.0
django-storages==1.14.2
django-stubs==4.2.7
django-stubs-ext==4.2.7
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.14.5
djlint==1.34.1
docutils==0.20.1
drf-spectacular==0.27.1
drf-yasg==1.21.7
EditorConfig==0.12.4
executing==2.0.1
factory-boy==3.3.0
Faker==23.2.1
filelock==3.13.1
future==1.0.0
gunicorn==21.2.0
hiredis==2.3.2
html-tag-names==0.1.2
html-void-elements==0.1.0
identify==2.5.35
idna==3.6
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.9.17
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
livereload==2.6.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mypy==1.7.1
mypy-extensions==1.0.0
netmiko==4.3.0
nodeenv==1.8.0
ntc_templates==4.3.0
oauthlib==3.2.2
packaging==23.2
paramiko==3.4.0
parso==0.8.3
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg-binary==3.1.18
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyserial==3.5
pytest==8.0.1
pytest-django==4.8.0
pytest-sugar==1.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
ruff==0.2.2
s3transfer==0.10.0
scp==0.14.5
sentry-sdk==1.40.5
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-autobuild==2024.2.4
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
stack-data==0.6.3
termcolor==2.4.0
text-unidecode==1.3
textfsm==1.1.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240218
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
virtualenv==20.25.1
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
Werkzeug==3.0.1
